requirements.txt
skale.py==3.8dev2
docker==4.2.0
pingparsing==1.0.3
schedule==0.6.0
tenacity==6.2.0
peewee==3.13.3
PyMySQL==0.9.3
